Only assign a point for a token when it is printed
A point could be assigned when a string containing a token was evaluated
without necessarily being printed. This happened in particular when
deciding the optimal direct object for an ambiguous GET command in the
presence of the string of pearls. Even if “string” was never visible on
screen, TADS was still evaluating its name string and assigning a point.

Now, token points are handled with an output filter.