Add a test emitting almost every string
This test is meant to output as much of the game’s text as practical
with the least number of input lines. This makes spell-checking easier.
Impractical strings are the entire source code and related messages from
PRAY TO IRIS, randomized strings e.g. from X ME, and the many unmodified
strings from the standard library. The run is unoptimized.