Add cover art
The cover art is based on the picture in the second column of folio 149
recto of MS. Douce 195. It is rotated 0.5° clockwise and cropped. The
text added below it uses the shades of red and blue found in that
manuscript’s line-terminal decorative fillers.

Because the original photograph is licensed under CC BY-NC 4.0 by
Bodleian Libraries, University of Oxford, the executable image must be
distributed accordingly, which I think means it can’t be sold. That is
unfortunate, but I wasn’t planning to sell it anyway, so it doesn’t
really matter.